commits
mkdir wide_n_deep_serving_client dir
